Functional stampfiles
Also redo our scraper so that we don't need to make a HEAD request
for each file; this will be much much faster in incremental ingest
cases. Due to some badness of apache mod_autoindex, we sacrifice
time resolution from second down to minute, but for our use cases
that shouldn't make a difference.