Basic usage of the girder python client library
This script has examples of editing images into large images, moving
items, deleting items and folders and adding meta data to items and
folders.This script assumes the large image plugin is enabled and
on. This script also assumes that users have authorization to delete
items and folders within the specified collection.