Fix brand name and banner color.
Recent changes to Girder (PR #2334) added a banner color, which broke how the banner looked when this plugin is enabled.  Girder also natively supports a brand name.  Use the core settings instead of adding them in this plugin.  This both reduces the plugin code and fixes an issue with the banner color.