Reinstall slicer_cli_web in provisioning.
As a convenience to developers, reinstall slicer_cli_web as part of
provisioning.  This makes it easier to bind-mount an external repo as
part of deployment (as is done for HistomicsUI in the devops scripts).

This also modifies the `devops/deploy.sh` script to only mount
directories that exists, and to mount both HistomicsUI and
slicer_cli_web as available or specified.