Mongo 7 upgrade
As part of our provisioning, we automatically ask mongo to upgrade the feature compatibility version.  This means that if you upgrade the DSA deployment using standard methods, it automatically upgrades your Mongo database as you go.  In theory, should one ever need to go back, you could downgrade the mongodb version manually a step at a time.  In practice, I'm not sure any deployments downgrade mongo.

As of Mongo 7, you can no longer undo upgrades to the DB.  The upgrade process requires an additional parameter because of this, so we aren't currently (2023-08-18) upgrade the feature compatibility version.  This will lead to problems on the next mongo release.

We have a flag in the provisioning (`mongo-compat`) that, if False, we
skip the compatibility version adjustment.

The actual downgrade path is probably to do a mongodump, remove the
mongo database, start with an older version of mongo, and do a
mongorestore.